Claude/fix deploy check 404 cq jvh

+++ flaskapp/migrations/add_location_based_budgeting.sql
@@ -164,8 +164,8 @@ CREATE TABLE IF NOT EXISTS campaign_location_targets (
     id INT AUTO_INCREMENT PRIMARY KEY,
 
     -- Campaign
-    campaign_id INT NOT NULL,
-    account_id INT NOT NULL,
+    campaign_id BIGINT NOT NULL,
+    account_id BIGINT NOT NULL,
 
     -- Location targeting
     location_criterion_id VARCHAR(50) NOT NULL COMMENT 'Google Ads location criterion ID',
